+++ packages/core/provider/src/Provider.tsx
@@ -14,8 +14,7 @@ export const NSUIProvider = ({ children, components, global }: NSUIProviderProps
   const contextValue: NSUIContextProps = {
     global: {
       motionDurations: {
-        color: global?.motionDurations?.color ?? 300,
-        scale: global?.motionDurations?.scale ?? 300,
+        fade: global?.motionDurations?.fade ?? 300,
         pop: global?.motionDurations?.pop ?? 600,
         ripple: global?.motionDurations?.ripple ?? 600
       },
@@ -24,6 +23,9 @@ export const NSUIProvider = ({ children, components, global }: NSUIProviderProps
     components: {
       button: {
         isRippleDisabled: components?.button?.isRippleDisabled ?? false
+      },
+      iconButton: {
+        isRippleDisabled: components?.iconButton?.isRippleDisabled ?? true
       },
       checkbox: {
         isPopDisabled: components?.checkbox?.isPopDisabled ?? false