+++ packages/ui/action/button/src/ButtonContext.ts
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export type ButtonContextValue = {
+  disabled?: boolean
+}
+
+export const ButtonContext = React.createContext<ButtonContextValue | null>(null)