+++ packages/ui/data-display/list/src/List.tsx
@@ -1,6 +1,7 @@
-import React, { forwardRef, ReactNode } from 'react'
+import React from 'react'
 import { cn, useNSUI } from '@negative-space/system'
 import { Flex, type FlexProps } from '@negative-space/flex'
+import { ListItem, type ListItemProps } from './ListItem'
 
 type ListElement = 'ol' | 'ul'
 
@@ -8,6 +9,8 @@ type ListDomMap = {
   ol: HTMLOListElement
   ul: HTMLUListElement
 }
+
+export type ListMarker = UnorderedMarker | OrderedMarker
 
 type UnorderedMarker = 'disc' | 'circle' | 'square'
 type OrderedMarker =
@@ -20,43 +23,60 @@ type OrderedMarker =
 
 type MarkerByElement<E extends ListElement> = E extends 'ol' ? OrderedMarker : UnorderedMarker
 
-type Direction = 'vertical' | 'horizontal'
-
 export type ListProps<E extends ListElement = 'ol'> = {
+  classNames?: {
+    root?: string
+    item?: string
+  }
+  styles?: {
+    root?: React.CSSProperties
+    item?: React.CSSProperties
+  }
   as?: E
-  children: ReactNode
+  items: ListItemProps[]
   marker?: MarkerByElement<E>
-  direction?: Direction
-} & Omit<FlexProps, 'as'> &
-  React.ComponentPropsWithoutRef<E>
+} & Omit<FlexProps, 'as' | 'children' | 'className' | 'style'> &
+  Omit<React.ComponentPropsWithoutRef<E>, 'className' | 'style'>
 
-export const List = forwardRef(
+export const List = React.forwardRef(
   <E extends ListElement = 'ol'>(
-    { as, children, className, marker, direction, ...props }: ListProps<E>,
+    { classNames, styles, as, items, marker, direction = 'column', ...props }: ListProps<E>,
     ref: React.Ref<ListDomMap[E]>
   ) => {
     const { global, components } = useNSUI()
     const listElement = as ?? components.list.typeElement
-    const Component = direction === 'horizontal' ? 'ul' : listElement
+    const Component = direction === 'row' ? 'ul' : listElement
 
-    const Marker =
-      direction === 'horizontal'
+    const Marker: ListMarker | 'none' =
+      direction === 'row'
         ? 'none'
-        : (marker ?? (listElement === 'ol' ? components.list.olMarker : components.list.ulMarker))
+        : (marker ??
+          (listElement === 'ol' ? components.list.olMarker : components.list.ulMarker) ??
+          'none')
 
     return (
       <Flex
         {...props}
         ref={ref}
         as={Component}
-        direction={direction === 'horizontal' ? 'row' : 'column'}
-        className={cn(
-          `${global.prefixCls}-list
-          ${direction !== 'horizontal' ? `${global.prefixCls}-marker-${Marker}` : ''}`,
-          className
-        )}
+        direction={direction}
+        className={cn(`${global.prefixCls}-list`, classNames?.root)}
+        style={{
+          listStyleType: Marker,
+          ...styles?.root
+        }}
       >
-        {children}
+        {items.map((item, index) => {
+          const key = `item-${index}`
+          return (
+            <ListItem
+              key={key}
+              {...(classNames?.item && { classNames: classNames.item })}
+              {...(styles?.item && { styles: styles.item })}
+              {...item}
+            />
+          )
+        })}
       </Flex>
     )
   }