+++ packages/ui/data-entry/radio/src/index.ts
@@ -1,2 +1,1 @@
-export { Radio, type RadioProps } from './Radio'
 export { RadioGroup, type RadioGroupProps } from './RadioGroup'