+++ packages/ui/layout/divider/__stories__/Divider.stories.tsx
@@ -6,8 +6,12 @@ export default {
   component: Divider,
   tags: ['autodocs'],
   args: {
-    className: 'w-md border-1 border-neutral-300'
+    className: 'w-md h-0.5 bg-neutral-300'
   }
 }
 
 export const Default = (args: DividerProps) => <Divider {...args} />
+
+export const Vertical = (args: DividerProps) => (
+  <Divider {...args} orientation="vertical" className="h-70 w-0.5 bg-neutral-300" />
+)