+++ packages/core/hooks/roving-focus/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @negative-space/roving-focus
 
+## 1.1.0
+
+### Minor Changes
+
+- 44eb1a5: Fix focus bug and add key handlers
+
 ## 1.0.0
 
 ### Major Changes