+++ packages/kits/nav/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @negative-space/nav
 
+## 1.0.6
+
+### Patch Changes
+
+- Updated dependencies [d4eb6c0]
+  - @negative-space/link@1.1.3
+
 ## 1.0.5
 
 ### Patch Changes