+++ packages/ui/feedback/spinner/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @negative-space/spinner
 
+## 2.1.3
+
+### Patch Changes
+
+- d4eb6c0: Change classNames
+- Updated dependencies [cd52e70]
+  - @negative-space/system@1.1.0
+
 ## 2.1.2
 
 ### Patch Changes