+++ packages/ui/typography/heading/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @negative-space/heading
 
+## 1.1.3
+
+### Patch Changes
+
+- d4eb6c0: Change classNames
+- Updated dependencies [cd52e70]
+  - @negative-space/system@1.1.0
+
 ## 1.1.2
 
 ### Patch Changes